+++ src/main/java/de/srendi/advancedperipherals/common/addons/appliedenergistics/AppEngApi.java
@@ -13,10 +13,10 @@ import dan200.computercraft.shared.util.NBTUtil;
 import de.srendi.advancedperipherals.AdvancedPeripherals;
 import de.srendi.advancedperipherals.common.util.LuaConverter;
 import de.srendi.advancedperipherals.common.util.Pair;
+import de.srendi.advancedperipherals.common.util.StringUtil;
 import it.unimi.dsi.fastutil.objects.Object2LongMap;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.world.item.ItemStack;
-import org.apache.commons.codec.binary.Hex;
 import org.apache.logging.log4j.Level;
 
 import java.nio.charset.StandardCharsets;
@@ -183,7 +183,7 @@ public class AppEngApi {
         try {
             byte[] bytesOfHash = fingerprint.getBytes(StandardCharsets.UTF_8);
             MessageDigest md = MessageDigest.getInstance("MD5");
-            return new String(Hex.encodeHex(md.digest(bytesOfHash)));
+            return StringUtil.toHexString(md.digest(bytesOfHash));
         } catch (NoSuchAlgorithmException ex) {
             AdvancedPeripherals.debug("Could not parse fingerprint.", Level.ERROR);
             ex.printStackTrace();