+++ src/main/java/de/srendi/advancedperipherals/common/blocks/tileentity/RsBridgeTile.java
@@ -1,0 +1,79 @@
+package de.srendi.advancedperipherals.common.blocks.tileentity;
+
+import com.refinedmods.refinedstorage.api.network.node.INetworkNodeProxy;
+import com.refinedmods.refinedstorage.blockentity.NetworkNodeBlockEntity;
+import com.refinedmods.refinedstorage.blockentity.config.IRedstoneConfigurable;
+import dan200.computercraft.api.peripheral.IPeripheral;
+import de.srendi.advancedperipherals.AdvancedPeripherals;
+import de.srendi.advancedperipherals.common.addons.computercraft.peripheral.RsBridgePeripheral;
+import de.srendi.advancedperipherals.common.addons.refinedstorage.RefinedStorageNode;
+import de.srendi.advancedperipherals.common.setup.TileEntityTypes;
+import de.srendi.advancedperipherals.lib.peripherals.IPeripheralTileEntity;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraftforge.common.capabilities.Capability;
+import net.minecraftforge.common.util.LazyOptional;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+import static dan200.computercraft.shared.Capabilities.CAPABILITY_PERIPHERAL;
+
+public class RsBridgeTile extends NetworkNodeBlockEntity<RefinedStorageNode> implements INetworkNodeProxy<RefinedStorageNode>, IRedstoneConfigurable, IPeripheralTileEntity {
+
+    private static final String PERIPHERAL_SETTINGS = "AP_SETTINGS";
+
+    protected CompoundTag peripheralSettings;
+
+    protected RsBridgePeripheral peripheral = new RsBridgePeripheral(this);
+    private LazyOptional<IPeripheral> peripheralCap;
+
+    public RsBridgeTile(BlockPos pos, BlockState state) {
+        super(TileEntityTypes.RS_BRIDGE.get(), pos, state);
+        peripheralSettings = new CompoundTag();
+    }
+
+    @NotNull
+    public <T1> LazyOptional<T1> getCapability(@NotNull Capability<T1> cap, @Nullable Direction direction) {
+        if (cap == CAPABILITY_PERIPHERAL) {
+            if (peripheral.isEnabled()) {
+                if (peripheralCap == null) {
+                    peripheralCap = LazyOptional.of(() -> peripheral);
+                }
+                return peripheralCap.cast();
+            } else {
+                AdvancedPeripherals.debug(peripheral.getType() + " is disabled, you can enable it in the Configuration.");
+            }
+        }
+        return super.getCapability(cap, direction);
+    }
+
+    public RefinedStorageNode createNode(Level level, BlockPos blockPos) {
+        return new RefinedStorageNode(level, blockPos);
+    }
+
+    @Override
+    public @NotNull void saveAdditional(@NotNull CompoundTag compound) {
+        super.saveAdditional(compound);
+        if (!peripheralSettings.isEmpty())
+            compound.put(PERIPHERAL_SETTINGS, peripheralSettings);
+    }
+
+    @Override
+    public void load(@NotNull CompoundTag compound) {
+        peripheralSettings = compound.getCompound(PERIPHERAL_SETTINGS);
+        super.load(compound);
+    }
+
+    @Override
+    public CompoundTag getPeripheralSettings() {
+        return peripheralSettings;
+    }
+
+    @Override
+    public void markSettingsChanged() {
+        setChanged();
+    }
+}