+++ CHANGELOG.md
@@ -14,6 +14,7 @@ We'll start adding more mod integrations in the next major 1.18 version.
 - [#247] We changed the textures of our blocks. Every block is now rotatable and the front texture of every block is now marked with a light aqua box.
 
 ### Added
+- Ported RsBridge peripheral from 1.16 to 1.18
 - [#237] Added `sendFormattedMessageToPlayer` to the chat box
 - [#237] Added the ability to format prefixes with the chat box
 - [#237] UUID's can now be used for `ToPlayer` functions in the chat box.